Removed top-level header to id user and logout since now regular app continues handling that. Removed connection form since server settings will not be exposed, and user/password are handled by new auth.
