+++ diplomacy/web/src/gui/pages/content_connection.jsx
@@ -16,7 +16,6 @@
 // ==============================================================================
 import React from 'react';
 import {Connection} from "../../diplomacy/client/connection";
-import {ConnectionForm} from "../forms/connection_form";
 import {DipStorage} from "../utils/dipStorage";
 import {Helmet} from "react-helmet";
 import {Navigation} from "../components/navigation";
@@ -24,29 +23,64 @@ import {PageContext} from "../components/page_context";
 import PropTypes from "prop-types";
 
 export class ContentConnection extends React.Component {
+
     constructor(props) {
         super(props);
         this.connection = null;
-        this.onSubmit = this.onSubmit.bind(this);
+        this.connect = this.connect.bind(this);
+
+        // Load fields values from local storage.
+        const initialState = this.initState();
+
+        const savedState = DipStorage.getConnectionForm();
+
+        if (savedState) {
+            if (savedState.hostname)
+                initialState.hostname = savedState.hostname;
+            if (savedState.port)
+                initialState.port = savedState.port;
+            if (savedState.username)
+                initialState.username = savedState.username;
+            if (savedState.showServerFields)
+                initialState.showServerFields = savedState.showServerFields;
+        }
+        this.state = initialState;
     }
 
-    onSubmit(data) {
+    initState() {
+        const { user } = this.props;
+
+        return {
+            hostname: window.location.hostname,
+            port: 8432,
+            username: user.username,
+            password: user.attributes.sub,
+            showServerFields: false
+        };
+    }
+
+    connect() {
+
+        const data = this.state;
         const page = this.context;
-        for (let fieldName of ['hostname', 'port', 'username', 'password', 'showServerFields'])
-            if (!data.hasOwnProperty(fieldName))
-                return page.error(`Missing ${fieldName}, got ${JSON.stringify(data)}`);
+
         page.info('Connecting ...');
+
         if (this.connection) {
             this.connection.currentConnectionProcessing.stop();
         }
+
         this.connection = new Connection(data.hostname, data.port, window.location.protocol.toLowerCase() === 'https:');
         this.connection.onReconnectionError = page.onReconnectionError;
+
         // Page is passed as logger object (with methods info(), error(), success()) when connecting.
         this.connection.connect(page)
             .then(() => {
                 page.connection = this.connection;
                 this.connection = null;
+
                 page.success(`Successfully connected to server ${data.username}:${data.port}`);
+
                 page.connection.authenticate(data.username, data.password)
                     .then((channel) => {
                         page.channel = channel;
@@ -80,24 +114,20 @@ export class ContentConnection extends React.Component {
     }
 
     render() {
-        const title = 'Connection';
+        const title = 'Connecting...';
         return (
             <main>
                 <Helmet>
                     <title>{title} | Diplomacy</title>
                 </Helmet>
                 <Navigation title={title}/>
-
-                <ConnectionForm
-                  onSubmit={this.onSubmit}
-                  user={this.props.user} />
-
             </main>
         );
     }
 
     componentDidMount() {
         window.scrollTo(0, 0);
+        this.connect();
     }
 }
